Create README.md
I conceptualized the idea of a late food truck night eatery in the Columbia area near Missou providing a cashless application for efficiency and speed.  I created all recipes, the menu, and the design of the application which will let customers pay on their cellphones and be notified when food is ready for pick up or 2 minutes prior to delivery.  The recorded metrics will promote efficiency since employees will be rewarded with an increase in hourly pay based on productivity.  It will also allow remote monitoring of metrics, live video, and sound.